Create Fractions
Fractions have only 2 variables:A numerator and a denominator. I chose to have only fractional values for calculation because it avoids rounding errors that can occur with decimals, and none of my matrix operations will ever require non rational (able to be expressed as a fraction) values, if given rational input.